requirements.txt
pyannote.audio
seaborn
tqdm
easydict